@@ -20,7 +20,7 @@ import love.disaster.j2j.core.spec.RemovrCompositeSpec;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
+import jakarta.inject.Inject;
 import java.util.HashMap;
 import java.util.Map;
 
