@@ -24,7 +24,7 @@ import love.disaster.j2j.core.shiftr.spec.ShiftrCompositeSpec;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
+import jakarta.inject.Inject;
 import java.util.HashMap;
 import java.util.Map;
 
